backend/requirements/dev.txt
-r base.txt

# Debug
django-debug-toolbar==4.4.6
ipython==8.30.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1